Refactor tests for poetry
