--- .github/workflows/actions.yml
@@ -1,83 +1,0 @@
-name: Actions
-
-on:
-  pull_request:
-    branches:
-      - master
-  push:
-    branches:
-      - master
-
-jobs:
-  lint:
-    name: Lint
-    runs-on: ubuntu-latest
-    steps:
-      - name: 📥 Checkout the repository
-        uses: actions/checkout@v2
-
-      - name: 🛠 Set up Python 3
-        uses: actions/setup-python@v2.3.1
-        id: python
-        with:
-          python-version: 3.x
-
-      - name: 📦 Install dependencies
-        run:  make requirements
-
-      - name: 🖤 Format with black
-        run:  black . --check
-
-      - name: 🗃 Organize with isort
-        run:  isort .
-
-      - name: 🧹 Lint files with pylint
-        run:  pylint pyefergy tests
-
-      - name: 🧹 Lint files with flake8
-        run:  flake8 pyefergy tests
-
-      - name: 🔍 Inspect with mypy
-        run: mypy pyefergy
-  test:
-    name: Test with Python ${{ matrix.python-version }}
-    needs: lint
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [3.8, 3.9, "3.10"]
-    steps:
-      - name: 📥 Checkout the repository
-        uses: actions/checkout@v2
-
-      - name: 🛠️ Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v2.3.1
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: 📦 Install dependencies
-        run:  make requirements
-
-      - name: 🏃 Run tests
-        run:  make test
-
-  coverage:
-    name: Upload coverage to Codecov
-    needs: test
-    runs-on: ubuntu-latest
-    steps:
-      - name: 📥 Checkout the repository
-        uses: actions/checkout@v2
-
-      - name: 🛠️ Set up Python
-        uses: actions/setup-python@v2.3.1
-        with:
-          python-version: 3.x
-
-      - name: 📦 Install dependencies
-        run:  make requirements
-
-      - name: 📤 Upload coverage to Codecov
-        run: |
-          make coverage
-          curl -sfSL https://codecov.io/bash | bash -
