+++ .github/workflows/python-publish.yml
@@ -1,0 +1,23 @@
+# An action to build and publish python package to https://pypi.org/ using poetry https://github.com/sdispater/poetry
+# For more information see: https://github.com/marketplace/actions/publish-python-poetry-package
+
+name: Publish python package
+
+on:
+  push:
+    tags:
+      - "*.*.*"
+
+permissions:
+  contents: read
+
+jobs:
+  publish:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4.1.2
+      - name: Build and publish to pypi
+        uses: JRubics/poetry-publish@v1.16
+        with:
+          pypi_token: ${{ secrets.PYPI_TOKEN }}
+          plugins: "poetry-dynamic-versioning[plugin]"
