--- MANIFEST.in
@@ -1,3 +1,0 @@
-include LICENSE
-exclude tests/*
-exclude pyproject.toml
