--- Makefile
@@ -1,25 +1,0 @@
-.DEFAULT_GOAL := help
-
-help: ## Shows this help message
-	@printf "\033[1m%s\033[36m %s\033[32m %s\033[0m \n\n" "Development environment for" "pyefergy" "";
-	@awk 'BEGIN {FS = ":.*##";} /^[a-zA-Z_-]+:.*?##/ { printf " \033[36m make %-25s\033[0m %s\n", $$1, $$2 } /^##@/ { printf "\n\033[1m%s\033[0m\n", substr($$0, 5) } ' $(MAKEFILE_LIST);
-	@echo
-
-requirements: ## Install requirements
-	@python3 -m pip --disable-pip-version-check install -r requirements.txt
-
-test: ## Run all tests
-	@python3 -m pytest tests -rxf -x -v -l --cov=./ --cov-report=xml
-
-lint: ## Lint all files black
-	@isort .
-	@python3 -m black --fast .
-	@python3 -m pylint pyefergy tests
-	@python3 -m flake8 pyefergy tests
-	@python3 -m mypy pyefergy
-
-coverage: ## Check the coverage of the package
-	@python3 -m pytest tests --cov=pyefergy --cov-report term-missing -vv
-
-setup: ## Setup the package
-	@python3 setup.py develop
