+++ pyproject.toml
@@ -14,6 +14,7 @@ disable = [
 
 [tool.mypy]
 ignore_missing_imports = true
+no_implicit_optional = false
 
 [tool.isort]
 # https://github.com/PyCQA/isort/wiki/isort-Settings
@@ -28,3 +29,42 @@ forced_separate = [
     "tests",
 ]
 combine_as_imports = true
+
+[tool.poetry]
+name = "pyefergy"
+version = "0.0.0"
+description = "An API library for Efergy energy meters."
+authors = ["Robert Hillis <tkdrob4390@yahoo.com>"]
+license = "MIT"
+readme = "README.md"
+homepage = "https://github.com/tkdrob/pyefergy"
+repository = "https://github.com/tkdrob/pyefergy"
+keywords = ["pyefergy", "efergy", "energy"]
+include = ["pyefergy/py.typed"]
+
+[tool.poetry.dependencies]
+python = ">=3.8,<4"
+aiohttp = ">=3.6.1"
+iso4217 = ">=1.2.20150619"
+pytz = ">=2024.1"
+types-pytz = ">=2021.3.1"
+
+[tool.poetry.group.dev.dependencies]
+aresponses = ">=2.1.4"
+black = ">=21.11b1"
+mypy = ">=0.910"
+pylint = ">=3.1.0"
+pytest-cov = ">=3.0.0"
+pytest-asyncio = ">=0.16.0"
+pytest-freezer = ">=0.4.8"
+pytest-timeout = "^2.3.1"
+pytest = ">=6.2.4"
+
+[tool.poetry-dynamic-versioning]
+enable = true
+vcs = "git"
+style = "semver"
+
+[build-system]
+requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
+build-backend = "poetry_dynamic_versioning.backend"
