+++ pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+addopts = --asyncio-mode=strict --timeout=10 --cov=pyefergy --cov-report term-missing -vv
