--- requirements.txt
@@ -1,14 +1,0 @@
-aiohttp>=3.6.1,<4.0
-aresponses>=2.1.4
-black>=21.11b1
-isort>=5.10.1
-flake8>=4.0.1
-flake8-docstrings>=1.6.0
-iso4217>=1.2.20150619
-mypy>=0.910
-pylint>=2.12.1
-pytest-cov>=3.0.0
-pytest-asyncio>=0.16.0
-pytest>=6.2.4
-pytz
-types-pytz==2021.3.1
