--- setup.cfg
@@ -1,21 +1,0 @@
-[egg_info]
-tag_build = 
-tag_date = 0
-
-[flake8]
-exclude = .venv,.git,.tox,docs,venv,bin,lib,deps,build
-max-complexity = 25
-doctests = True
-# To work with Black
-# E501: line too long
-# W503: Line break occurred before a binary operator
-# E203: Whitespace before ':'
-# D202 No blank lines allowed after function docstring
-# W504 line break after binary operator
-ignore =
-    E501,
-    W503,
-    E203,
-    D202,
-    W504
-noqa-require-code = True
